Detection and selection of sound schemes
Automatically detect schemes to install at configure time,
in the same way as episodes.

Fallback from the current scheme to the default if the
current scheme does not have a sound file for a particular
event (Repton 1, for example, is sparse).

New sound 'Start' to play at start of game, fallback to the
original 'Transporter' if this is not present.